relocate to foliascheduler to avoid confusions
