+++ platforms/folia/src/main/java/com/cjcrafter/foliascheduler/folia/FoliaAsyncScheduler.java
@@ -1,26 +1,28 @@
-package com.cjcrafter.scheduler.folia;
+package com.cjcrafter.foliascheduler.folia;
 
-import com.cjcrafter.scheduler.SchedulerImplementation;
-import com.cjcrafter.scheduler.TaskImplementation;
-import io.papermc.paper.threadedregions.scheduler.GlobalRegionScheduler;
+
+import com.cjcrafter.foliascheduler.AsyncSchedulerImplementation;
+import com.cjcrafter.foliascheduler.TaskImplementation;
+import io.papermc.paper.threadedregions.scheduler.AsyncScheduler;
 import io.papermc.paper.threadedregions.scheduler.ScheduledTask;
 import org.bukkit.plugin.Plugin;
 import org.jetbrains.annotations.ApiStatus;
 import org.jetbrains.annotations.NotNull;
 
+import java.util.concurrent.TimeUnit;
 import java.util.function.Consumer;
 import java.util.function.Function;
 
 @ApiStatus.Internal
-public class FoliaGlobalScheduler implements SchedulerImplementation {
+public class FoliaAsyncScheduler implements AsyncSchedulerImplementation {
 
     private final @NotNull Plugin plugin;
-    private final @NotNull GlobalRegionScheduler globalRegionScheduler;
+    private final @NotNull AsyncScheduler asyncScheduler;
 
     @ApiStatus.Internal
-    public FoliaGlobalScheduler(@NotNull Plugin plugin) {
+    public FoliaAsyncScheduler(@NotNull Plugin plugin) {
         this.plugin = plugin;
-        this.globalRegionScheduler = plugin.getServer().getGlobalRegionScheduler();
+        this.asyncScheduler = plugin.getServer().getAsyncScheduler();
     }
 
     private <T> @NotNull Consumer<ScheduledTask> buildFoliaConsumer(
@@ -35,33 +37,28 @@ public class FoliaGlobalScheduler implements SchedulerImplementation {
     }
 
     @Override
-    public void execute(@NotNull Runnable run) {
-        globalRegionScheduler.execute(plugin, run);
-    }
-
-    @Override
-    public @NotNull <T> TaskImplementation<T> run(@NotNull Function<TaskImplementation<T>, T> function) {
+    public <T> @NotNull TaskImplementation<T> runNow(@NotNull Function<TaskImplementation<T>, T> function) {
         FoliaTask<T> taskImplementation = new FoliaTask<>();
         Consumer<ScheduledTask> foliaConsumer = buildFoliaConsumer(taskImplementation, function);
-        ScheduledTask scheduledTask = globalRegionScheduler.run(plugin, foliaConsumer);
+        ScheduledTask scheduledTask = asyncScheduler.runNow(plugin, foliaConsumer);
         taskImplementation.setScheduledTask(scheduledTask);
         return taskImplementation;
     }
 
     @Override
-    public @NotNull <T> TaskImplementation<T> runDelayed(@NotNull Function<TaskImplementation<T>, T> function, long delay) {
+    public <T> @NotNull TaskImplementation<T> runDelayed(@NotNull Function<TaskImplementation<T>, T> function, long delay, @NotNull TimeUnit unit) {
         FoliaTask<T> taskImplementation = new FoliaTask<>();
         Consumer<ScheduledTask> foliaConsumer = buildFoliaConsumer(taskImplementation, function);
-        ScheduledTask scheduledTask = globalRegionScheduler.runDelayed(plugin, foliaConsumer, delay);
+        ScheduledTask scheduledTask = asyncScheduler.runDelayed(plugin, foliaConsumer, delay, unit);
         taskImplementation.setScheduledTask(scheduledTask);
         return taskImplementation;
     }
 
     @Override
-    public @NotNull <T> TaskImplementation<T> runAtFixedRate(@NotNull Function<TaskImplementation<T>, T> function, long delay, long period) {
+    public <T> @NotNull TaskImplementation<T> runAtFixedRate(@NotNull Function<TaskImplementation<T>, T> function, long delay, long period, @NotNull TimeUnit unit) {
         FoliaTask<T> taskImplementation = new FoliaTask<>();
         Consumer<ScheduledTask> foliaConsumer = buildFoliaConsumer(taskImplementation, function);
-        ScheduledTask scheduledTask = globalRegionScheduler.runAtFixedRate(plugin, foliaConsumer, delay, period);
+        ScheduledTask scheduledTask = asyncScheduler.runAtFixedRate(plugin, foliaConsumer, delay, period, unit);
         taskImplementation.setScheduledTask(scheduledTask);
         return taskImplementation;
     }
