+++ src/main/java/com/cjcrafter/foliascheduler/ServerImplementation.java
@@ -1,4 +1,4 @@
-package com.cjcrafter.scheduler;
+package com.cjcrafter.foliascheduler;
 
 import org.bukkit.Chunk;
 import org.bukkit.Location;
